Indexing subfolders for easier component imports
