+++ assets/views/AboutPage/index.js
@@ -1,0 +1,3 @@
+import AboutPage from "./AboutPage";
+
+export default AboutPage;
