+++ assets/views/ContactPage/index.js
@@ -1,0 +1,3 @@
+import ContactPage from "./ContactPage";
+
+export default ContactPage;
