+++ assets/views/LandingPage/index.js
@@ -1,0 +1,3 @@
+import LandingPage from "./LandingPage";
+
+export default LandingPage;
