+++ assets/views/PortfolioPage/index.js
@@ -1,0 +1,3 @@
+import PortfolioPage from "./PortfolioPage";
+
+export default PortfolioPage;
