+++ assets/views/ResumePage/index.js
@@ -1,0 +1,3 @@
+import ResumePage from "./ResumePage";
+
+export default ResumePage;
